+++ automax/automaxcpu_test.go
@@ -8,8 +8,9 @@ import (
 	"go.akshayshah.org/attest"
 )
 
-func TestSetCpu(t *testing.T) {
+func TestSetCpu(t *testing.T) { //nolint:tparallel
 	t.Parallel()
+	// This tests can run in parallel with others but not with themselves.
 
 	write := func(cgroupV2Value string) string {
 		dir := t.TempDir()
@@ -27,8 +28,6 @@ func TestSetCpu(t *testing.T) {
 	}
 
 	t.Run("cgroupV2", func(t *testing.T) {
-		t.Parallel()
-
 		/*
 			use 3 cpus.
 			docker run -it --entrypoint /bin/bash --cpus="3" redis
@@ -48,8 +47,6 @@ func TestSetCpu(t *testing.T) {
 	})
 
 	t.Run("cgroup max", func(t *testing.T) {
-		t.Parallel()
-
 		fName := write("max 100000")
 
 		expected := currentMaxProcs()
@@ -64,8 +61,6 @@ func TestSetCpu(t *testing.T) {
 	})
 
 	t.Run("cpu less than 1", func(t *testing.T) {
-		t.Parallel()
-
 		/*
 			use 50% of cpu.
 			docker run -it --entrypoint /bin/bash --cpus=".5" redis
@@ -85,8 +80,6 @@ func TestSetCpu(t *testing.T) {
 	})
 
 	t.Run("one field", func(t *testing.T) {
-		t.Parallel()
-
 		fName := write("500000")
 
 		expected := int(5)
@@ -102,8 +95,6 @@ func TestSetCpu(t *testing.T) {
 	})
 
 	t.Run("corrupt file", func(t *testing.T) {
-		t.Parallel()
-
 		fName := write("its corrupt")
 
 		expected := currentMaxProcs()